Move utils from use-onboarding-sections to own files
